[wip] Implement deep ETags for Facia
If the client sends an If-None-Match header for a Facia page, we can use
it to check whether we need to bother deserializing the (potentially
large) json file retrieved from S3. If the received hash matches the MD5
of the raw json, we know that the client already has the latest version,
so we can return 304 Not Modified. Otherwise we proceed as normal.